PS2: fix oppo redeclaration

+++ media/libstagefright/MediaCodecList.cpp
@@ -65,7 +65,6 @@ MediaCodecList::MediaCodecList()
 
         addMediaCodec(true /* encoder */, "AACEncoder", "audio/mp4a-latm");
 
-
         addMediaCodec(
                      false /* encoder */, "OMX.google.raw.decoder", "audio/raw");
 
@@ -85,7 +84,7 @@ MediaCodecList::MediaCodecList()
         AString line = info.mName;
         line.append(" supports ");
         for (size_t j = 0; j < mTypes.size(); ++j) {
-            uint32_t value = mTypes.valueAt(j);
+            uint64_t value = mTypes.valueAt(j);
 
             if (info.mTypes & (1ul << value)) {
                 line.append(mTypes.keyAt(j));
@@ -403,7 +402,7 @@ status_t MediaCodecList::addTypeFromAttributes(const char **attrs) {
 }
 
 void MediaCodecList::addType(const char *name) {
-    uint32_t bit;
+    uint64_t bit;
     ssize_t index = mTypes.indexOfKey(name);
     if (index < 0) {
         bit = mTypes.size();
@@ -430,7 +429,7 @@ ssize_t MediaCodecList::findCodecByType(
         return -ENOENT;
     }
 
-    uint32_t typeMask = 1ull << mTypes.valueAt(typeIndex);
+    uint64_t typeMask = 1ull << mTypes.valueAt(typeIndex);
 
     while (startIndex < mCodecInfos.size()) {
         const CodecInfo &info = mCodecInfos.itemAt(startIndex);
